refactor: add CycleCard
